requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.42
werkzeug==3.1.3
gunicorn==23.0.0
discord.py==2.5.2
python-dotenv==1.1.1
apscheduler==3.11.0
pytz==2023.3
psutil==7.0.0